Created a gamepiece class and viewmodel.
Add add/ermove piece functions to gameboard that updates both viewmodels and models as needed.  Add enum to create default shapes for circles (ellipse) and squares in associated viewmodels.

Lots of refactoring out of models.  Cleaned up and condensed code of view models in preparation to use bindings for the gameboard in the UI.